web_api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
supabase>=2.0.0
